+++ generators/client-2/templates/src/main/webapp/app/account/_account.module.ts
@@ -44,12 +44,4 @@ angular
     <%_ if (authenticationType === 'session') { _%>
     .directive('sessions', <angular.IDirectiveFactory> upgradeAdapter.downgradeNg2Component(SessionsComponent))
     <%_ } _%>
-    .directive('settings', <angular.IDirectiveFactory> upgradeAdapter.downgradeNg2Component(SettingsComponent))
-    .factory('Register', upgradeAdapter.downgradeNg2Provider(Register))
-    .factory('Activate', upgradeAdapter.downgradeNg2Provider(Activate))
-    .factory('Password',upgradeAdapter.downgradeNg2Provider( Password))
-    .factory('PasswordResetInit', upgradeAdapter.downgradeNg2Provider(PasswordResetInit))
-    <%_ if (authenticationType === 'session') { _%>
-    .factory('SessionsService', upgradeAdapter.downgradeNg2Provider(SessionsService))
-    <%_ } _%>
-    .factory('PasswordResetFinish', upgradeAdapter.downgradeNg2Provider(PasswordResetFinish));
+    .directive('settings', <angular.IDirectiveFactory> upgradeAdapter.downgradeNg2Component(SettingsComponent));
